+++ tests/cases/TestApplication.php
@@ -1,0 +1,13 @@
+<?php
+
+use PHPUnit\Framework\TestCase;
+use ntentan\middleware\MiddlewareQueue;
+
+class TestApplication extends TestCase
+{
+    public function testSingleItemQueue()
+    {
+        $middlewareQueue = new MiddlewareQueue();
+
+    }
+}