+++ tests/cases/TestMiddlewareQueue.php
@@ -1,0 +1,13 @@
+<?php
+
+use PHPUnit\Framework\TestCase;
+
+class TestMiddlewareQueue extends TestCase
+{
+    public function setUp(): void
+    {
+
+    }
+
+    public function
+}